Move SVG files to app/assets/svgs and remove commented out test
