+++ .eslintignore
@@ -48,7 +48,7 @@ components/MultiSelect/multi-select-input.tsx
 components/Popover/popover.tsx
 components/ProfileCard/ProfileCard.tsx
 components/SnowOverlayWrapper/react-snow-overlay.d.ts
-components/LogoCombination/LogoCombination.tsx
-components/LogoIcon/LogoIcon.tsx
+app/assets/svgs/LogoCombination.tsx
+app/assets/svgs/LogoIcon.tsx
 lib/supabase/middleware.ts
 lib/utils.ts
