+++ app/assets/svgs/LogoCombination.tsx
