+++ app/assets/svgs/LogoIcon.tsx
