+++ components/NavLogo/NavLogo.test.tsx
@@ -5,16 +5,13 @@ import NavLogo from './NavLogo';
 import { render, screen } from '@testing-library/react';
 
 describe('NavLogo', () => {
-    it('renders the logo and links to the dashboard', () => {
-        render(<NavLogo/>);
+  it('renders the logo and links to the dashboard', () => {
+    render(<NavLogo />);
 
-        const navLogoIcon = screen.getByTestId('nav-logo-icon');
-        expect(navLogoIcon).toBeInTheDocument();
-        
-        // const navLogoText = screen.getByTestId('nav-logo-text');
-        // expect(navLogoText).toHaveTextContent('Elfgorithm');
+    const navLogoIcon = screen.getByTestId('nav-logo-icon');
+    expect(navLogoIcon).toBeInTheDocument();
 
-        const navLogoLink = screen.getByTestId('nav-logo-link');
-        expect(navLogoLink).toHaveAttribute('href', '/dashboard')
-    });
-})
+    const navLogoLink = screen.getByTestId('nav-logo-link');
+    expect(navLogoLink).toHaveAttribute('href', '/dashboard');
+  });
+});
